// src/models/authInterface.ts
export interface otpStrengthInterface {
    text: string;
    textColor: string;
    first: string;
    second: string;
    third: string;
  }


  export interface initialSignUpDataInterface {
    userName: string;
    userEmail: string;
    phoneNumber: string;
    dateOfBirth: string;
    gender: string;
  }
  